changed mprsync naming to sync-data; updated mprsync-root.sh -> sync-data.sh script
- renamed mprsync configs and mprsync-root.sh to sync-data naming because it uses
  mprsync.sh but is not related to it directly to be its configuration
- updated sync-data.sh script to apply diversions automatically
- updated sync-data.sh script to deal with changes in auto/manual marks on packages
- updated sync-data.sh script to automatically encrypt any newly recorded directories
- changed "#!/bin/bash -e" and "#!/bin/sh -e" to explicit "set -e" later so that it
  works both when invoked directly and using bash/sh (i.e. bash <script> or sh <script>)

+++ .local/bin/enc-init.sh
@@ -1,4 +1,6 @@
-#!/bin/sh -e
+#!/bin/sh
+
+set -e
 
 # Encrypt listed directories making a copy of those. This can be usually done while
 # logged into GUI (close all other apps to ensure files being copied are not being
@@ -8,7 +10,7 @@
 
 ENC_PREFIX="enc-init"
 for dir in `cat enc.dirs`; do
-  if [ -d $dir -a ! -e $dir.$ENC_PREFIX ]; then
+  if [ -d $dir -a ! -e $dir.$ENC_PREFIX ] && ! fscrypt status $dir >/dev/null 2>/dev/null; then
     mkdir $dir.$ENC_PREFIX && \
       fscrypt encrypt $dir.$ENC_PREFIX --no-recovery && \
       cp -a -T $dir $dir.$ENC_PREFIX
